Porting a lot of code to use single precision floating point numbers.
Rewrote base WPILib classes to support single precision floating operations. The `DriveTrainSubsytem` now uses these.

This needs to be tested ASAP, this could break a lot of code if not properly tested. Hypothetically it should be fine since all of the code was copied, but we really should do a lot of testing for it.