Various fixes.
Fixed some math (degrees to radians). Removed ambiguity from the logger (compiler error). Reverted more stuff to using double precision floating point numbers, if we use floats it will constantly have to up scale to doubles. The Cortex-A9 supports 32 bit and 64 bit floating point arithmetic, this is fairly vague, it could mean that it will do 2 operations on 64 bit numbers or that it will upscale 32 bit numbers to the full 64 bits, or that it has two different processors for 32 bit and 64 operations (this one is unlikely). We are left with two options, always use doubles to avoid upscaling, or convert *everything* to using floats. For now I am going to use doubles, if I find out that it does two 32 bit operations to perform a 64 bit operation I will switch the code to using single precision floating point number exclusively.