Change Pure Pursuit paths
A note... for some reason the paths seem to ned assymetric distances when testing on the test field... this is probably due to one side (left) having carpet and the other side (right) having tile
It would be amazing if we could figure out how to have the same movement for each side... however, we are not stopping at 0 ft/s which might be the reason it is being weird... this will make different surfaces
have a highly different effect.

Additionally, I implemented the ability for the robot to ONLY go straight in autonomous to avoid running into our alliance partners.

We also NEED to check if the robot log system works