(UNTESTED) Change same side scale auton to go to the left side of the scale
This will make our robot compatible with 2052's cross country scale autonomous because they come up on the near side of the scale, so we must be out of their way.
It is not guaranteed that we will be with 2052 in elims, however it is very likely that an alliance partner will have a cross country scale auto
These points came from the CAD model of the field and should be very close if not perfect.