Fix same side switch auto
We never know if we will need it. Currently, the program is innately biased towards the scale. Because of this, the SWITCH option in the strategy selector will only run the switch in case of LL or RR. In case of LR on the left or RL on the right, it will do a cross-country auton. It is trivial to allow the switch auton to be run in these situations as well.