"Smoother" motion profiling, backup motion profile for first cube
I have made it possible (and easy) to do most of the preprocessing for trajectory points (required to feed them into talons) on robot init. This means that there will no longer be pauses in between Motion Profiling Commands. Additionally, in case Pure Pursuit is RIP, I have added a backup motion profile for the first cube on the right side