Working towards Two-Cube
ISSUES: Elevator does not lift all the way to the top while driving during fast movements. This is likely a mechanical/electrical bug that needs to be resolved, but everyone is unsure as to what it could be. The elevator motors seem to be drawing far too much current. We have made many attempts to get around this via software (not driving fast and raising elevator at the same time).