Replace lookahead circle with text
This is because circle-line intersection is not how PP finds its goal
point. Instead, there is a length of string that is 1 lookahead long in
front of the robot. It will try to wrap it around the upcoming path
segments. So if the robot is driving on a straight length of segment, it
will have the same output as circle line intersection. However, it it
encounters a curve, the point will be closer.