Add Spline Waypoints and Spline Paths
This makes it possible to use cubic parametric splines as path segments using the SplinePathSegment and SplinePath class

Naturally, we are able to set the angle of the spline at a control point. Right now, the tangent vector is scaled based on the max velocity. Is that what we want?

Also, all the parametric arcLength functions are numerically based because an online integral calculator told me that the indefinite integral for finding an the arc length of a cubic parametric does not exist