Splines no longer instantly crash
The issue was in Path#fromSplinePoints, where I interpolate maxVel, maxAccel, and maxDecel. I was putting everything in the Interpolationmap for maxVel, which led to some waypoints
having a negative max speed. Of course, this led to the robot getting stuck at the point where the waypoints went from positive maxVel to negative maxVel.